fix for [issue #8](https://github.com/ElhamAryanpur/BlueEngine/issues/8)
Issue: Blue Engine not compiling for MacOS because of NON_FILL_POLYGON_MODE feature.

NON_FILL_POLYGON_MODE helps with points/line only and wireframes of objects, although it's only support on DX12 and Vulkan as of yet.
So the solution was to check the target system, and disable it for MacOS.